requirements.txt
PyQt6
numpy
matplotlib
pytest
toml
PyOpenGL
PyOpenGL_accelerate